requirements.txt
Django==3.0.8
jinjalint==0.5
42cc-pystyle
black==20.8b1
flake8==3.8.3
flake8-black==0.2.1
flake8-broken-line==0.2.1
flake8-cognitive-complexity==0.1.0
flake8-commas==2.0.0
flake8-django==1.1.1
flake8-eradicate==0.4.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-print==3.1.4
flake8-quotes==3.2.0
flake8-variables-names==0.0.3
django-manifest-loader==1.0.0
djangorestframework==3.12.2
